refactor(match): move shared overlay types [GHI#101]
