+++ src/ui/mobile/modules/match/components/match-result/MatchResultOverlay.tsx
@@ -1,6 +1,6 @@
+import type { MatchResultOverlayProps } from "@/ui/shared/match/types/matchResultOverlay";
 import { getMatchResultViewModel } from "@/ui/shared/match/utils";
 
-import type { MatchResultOverlayProps } from "./MatchResultOverlay.types";
 import { MatchResultOverlayBase } from "./MatchResultOverlayBase";
 
 export const MatchResultOverlay = ({
